+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "external/raylib"]
+	path = external/raylib
+	url = https://github.com/raysan5/raylib
