+++ examples/resources/clips/LICENSE
@@ -1,0 +1,5 @@
+LICENSE
+
+The video clips in this folder are sourced from Hikingfex Free Video (https://www.hikingfex.com/free-video). Each clip is labeled as 100% royalty-free and no copyright by Hikingfex, indicating they are free to use without any copyright restrictions or licensing fees.
+
+License: No copyright, royalty-free, as stated by the original source.
